chore: Refactor l1 tx utils (#17275)
Split l1-tx-utils into several files before going into further changes.
