+++ yarn-project/ethereum/src/contracts/empire_base.ts
@@ -4,7 +4,7 @@ import { EmpireBaseAbi } from '@aztec/l1-artifacts/EmpireBaseAbi';
 
 import { type Hex, type TypedDataDefinition, encodeFunctionData } from 'viem';
 
-import type { L1TxRequest } from '../l1_tx_utils.js';
+import type { L1TxRequest } from '../l1_tx_utils/index.js';
 
 export interface IEmpireBase {
   get address(): EthAddress;
