+++ yarn-project/ethereum/src/contracts/fee_asset_handler.ts
@@ -3,7 +3,7 @@ import { FeeAssetHandlerAbi } from '@aztec/l1-artifacts/FeeAssetHandlerAbi';
 
 import { type Hex, encodeFunctionData, getContract } from 'viem';
 
-import type { L1TxUtils } from '../l1_tx_utils.js';
+import type { L1TxUtils } from '../l1_tx_utils/index.js';
 
 export class FeeAssetHandlerContract {
   public address: EthAddress;
