+++ yarn-project/ethereum/src/contracts/multicall.test.ts
@@ -13,7 +13,7 @@ import { foundry } from 'viem/chains';
 import { createExtendedL1Client } from '../client.js';
 import { DefaultL1ContractsConfig } from '../config.js';
 import { type DeployL1ContractsReturnType, deployL1Contract, deployL1Contracts } from '../deploy_l1_contracts.js';
-import { L1TxUtils, createL1TxUtilsFromViemWallet } from '../l1_tx_utils.js';
+import { L1TxUtils, createL1TxUtilsFromViemWallet } from '../l1_tx_utils/index.js';
 import { startAnvil } from '../test/start_anvil.js';
 import type { ExtendedViemWalletClient } from '../types.js';
 import { FormattedViemError } from '../utils.js';
