+++ yarn-project/ethereum/src/index.ts
@@ -1,7 +1,7 @@
 export * from './constants.js';
 export * from './deploy_l1_contracts.js';
 export * from './chain.js';
-export * from './l1_tx_utils.js';
+export * from './l1_tx_utils/index.js';
 export * from './l1_contract_addresses.js';
 export * from './l1_reader.js';
 export * from './utils.js';
