+++ yarn-project/ethereum/src/l1_tx_utils/index.ts
@@ -1,0 +1,11 @@
+export * from './config.js';
+export * from './constants.js';
+export * from './factory.js';
+export * from './l1_tx_utils.js';
+export * from './readonly_l1_tx_utils.js';
+export * from './signer.js';
+export * from './types.js';
+export * from './utils.js';
+
+// Note: We intentionally do not export l1_tx_utils_with_blobs.js
+// to avoid accidentally importing blob-lib dependency.
