+++ yarn-project/ethereum/src/test/upgrade_utils.ts
@@ -6,7 +6,7 @@ import { type GetContractReturnType, type PrivateKeyAccount, getContract } from 
 
 import { extractProposalIdFromLogs } from '../contracts/governance.js';
 import type { L1ContractAddresses } from '../l1_contract_addresses.js';
-import { createL1TxUtilsFromViemWallet } from '../l1_tx_utils.js';
+import { createL1TxUtilsFromViemWallet } from '../l1_tx_utils/index.js';
 import type { ExtendedViemWalletClient, ViemPublicClient } from '../types.js';
 import { EthCheatCodes } from './eth_cheat_codes.js';
